+++ NEWS.md
@@ -3,6 +3,8 @@
 * `dbListFields` now works with identifier name which in turn supports the use
   of `in_schema()` in `tbl()` (#200)
 * Fix an error whereby join query's CTEs are not properly retrieved (#203)
+* Fix a bug that causes NULL values in a ROW to return data-schema mismatch
+  error (#206)
 
 # RPresto 1.4.0